Module Context.Abi_v2

ABI V2 is the version 2 of the supported list of ABIs

type t =

| Android_arm64v8a

| Android_arm32v7a

| Android_x86

| Android_x86_64

| Darwin_arm64

| Darwin_x86_64

| Linux_arm64

| Linux_arm32v6

| Linux_arm32v7

| Linux_x86_64

| Linux_x86

| Windows_x86_64

| Windows_x86

| Windows_arm64

| Windows_arm32

val equal : t -> t -> Ppx_deriving_runtime.bool

val compare : t -> t -> Ppx_deriving_runtime.int

val min : int

val max : int

val to_enum : t -> int

val of_enum : int -> t option

val of_string : string -> (t, string) Stdlib.result

val to_string : t -> string

to_string abi is the enumeration value of abi; for example "Linux_x86".

val to_canonical_string : t -> string

to_canonical_string abi will give the canonical representation of the ABI to DKML tools and APIs.

val show : t -> string

val pp : Stdlib.Format.formatter -> t -> unit

val is_windows : t -> bool

val is_linux : t -> bool

val is_darwin : t -> bool

val is_android : t -> bool

val word_size : t -> int

word_size is the number of bits in a word for the machine's CPU architecture. The number of bits will be 32 or 64, but may be something else for more exotic future architectures.

val values : t list