Module Dkml_install_register.Component_registry

The Component_registry is a global registry of all components that have been registered until now.

Component authors should follow this sequence:

open Dkml_install_api

module Component : Component_config = struct
  include Default_component_config
  let component_name = "...the..component..name..."
  (** Redefine any other values you want to override *)
end
let reg = Component_registry.get ()
let () = Component_registry.add_component reg (module Component : Component_config)

type t

The type of the component registry

val get : unit -> t

Get a reference to the global component registry

val add_component : t -> (module Dkml_install_api.Component_config) -> unit

add_component registry component adds the component to the registry

val validate : t -> ( unit, string ) Stdlib.result

validate registry succeeds if and only if all dependencies of all add_component registry have been themselves added

val eval : t -> f: ( (module Dkml_install_api.Component_config) -> ( 'a, string ) Stdlib.result ) -> ( 'a list, string ) Stdlib.result

eval registry ~f iterates through the registry in dependency order, executing function f on each component configuration.

val reverse_eval : t -> f: ( (module Dkml_install_api.Component_config) -> ( 'a, string ) Stdlib.result ) -> ( 'a list, string ) Stdlib.result

reverse_eval registry ~f iterates through the registry in reverse dependency order, executing function f on each component configuration.

module Private : sig ... end

The module Private is meant for internal use only.