Module Dkml_install_api

Component Configuration

You are responsible for creating a module of type Component_config to describe your component. A do-nothing implementation for most of the module is available as Default_component_config.

Dkml_install_api will use the Component_config to create four (4) command line applications.

Each of the command line applications are "subcommands" in the language of the OCaml Cmdliner package. You will not need to understand Cmdliner to define your own component, although you may visit the Cmdliner documentation if you want more information.

The four (4) command line applications have limited access to the OCaml runtime. The expectation is that all installation logic is embedded in bytecode executables which have the complete set of package dependencies you need to run your logic. Through Component_config Dkml_install_api will have given you a ~ctx_t Cmdliner term that, when evaluated, leads to the context record Context.t. The context record has the information needed to run your bytecode executables.

On Windows it is recommended security practice to separate functionality that requires administrative privileges from functionality that does not require non-administrative privileges. Dkml_install_api follows the same recommendations:

There will be a single executable "dkml-install-admin-runner.exe" that is responsible for the following functionality for all components:

Administrator installation defined by Component_config.install_admin_subcommand

Administrator uninstallation defined by Component_config.uninstall_admin_subcommand

There will be a single executable "dkml-install-user-runner.exe" that is responsible for the following functionality for all components:

User installation defined by Component_config.install_user_subcommand

User uninstallation defined by Component_config.uninstall_user_subcommand

module Context : sig ... end

Context is a module providing a record type for the context.

module type Component_config_defaultable = sig ... end

Component configuration values that can be supplied with defaults.

module type Component_config = sig ... end

Each component must define a configuration module

You should include Default_component_config in any of your components so that your component can be future-proof against changes in the Component_config signature.

module Default_component_config : sig ... end

Default values for a subset of the module type Component_config.

module Log_config : module type of Log_config

exception Installation_error of string

Raise Installation_error message when your component has a terminal failure

val log_spawn_and_raise : Bos.Cmd.t -> unit

log_spawn_and_raise cmd logs the command cmd and runs it synchronously, raising Installation_error if the command exits with a non-zero error code.

The environment variable "OCAMLRUNPARAM" will be set to "b" so that any OCaml bytecode launched by log_spawn_and_raise will have backtraces. Any exiting environment variable "OCAMLRUNPARAM" will be kept, however.