Module Dkml_install_register.Component_registry

The Component_registry is a global registry of all components that have been registered until now.

Component authors should follow this sequence:

open Dkml_install_api

module Component : Component_config = struct
    (** Fill this in *)
end
let reg = Component_registry.get ()
let run = function Ok () -> () | Error str -> failwith str
let () = run
  @@ Component_registry.add_component reg (module Component : Component_config)

type t

The type of the component registry

val get : unit -> t

Get a reference to the global component registry

val add_component : t -> (module Dkml_install_api.Component_config) -> (unit, string) Stdlib.Result.t

add_component registry component adds the component to the registry

val iter : t -> f:(string -> (module Dkml_install_api.Component_config) -> unit) -> unit

iter registry ~f iterates through the registry, executing function f on each component name and configuration.